* 'main' of github.com:ttscoff/marked-conductor:
  add test tracks
  Version bump
  ### 1.0.28
  more tests more fixes
  more tests more fixes
  add tests
  Switch to using #size

+++ lib/conductor/config.rb
@@ -3,31 +3,60 @@
 module Conductor
   # Configuration methods
   class Config
-    attr_reader :config, :tracks
+    # Configuration
+    attr_reader :config
+    # Tracks element
+    attr_reader :tracks
+    # Config file path
+    attr_writer :config_file
 
+    ##
+    ## Instantiate a configuration
+    ##
+    ## @return     [Config] Config object
+    ##
     def initialize
-      config_file = File.expand_path("~/.config/conductor/tracks.yaml")
+      @config_file = File.expand_path("~/.config/conductor/tracks.yaml")
+    end
 
-      create_config(config_file) unless File.exist?(config_file)
+    def configure
+      res = create_config(@config_file)
+      return false unless res
 
-      @config ||= YAML.safe_load(IO.read(config_file))
+      @config ||= YAML.safe_load(IO.read(@config_file))
 
       @tracks = @config["tracks"].symbolize_keys
+
+      return true
     end
 
     private
 
-    def create_config(config_file)
+    ##
+    ## Generate a blank config and directory structure
+    ##
+    ## @param      config_file  [String] The configuration file to create
+    ##
+    def create_config(config_file = nil)
+      config_file ||= @config_file
       config_dir = File.dirname(config_file)
       scripts_dir = File.dirname(File.join(config_dir, "scripts"))
       FileUtils.mkdir_p(config_dir) unless File.directory?(config_dir)
       FileUtils.mkdir_p(scripts_dir) unless File.directory?(scripts_dir)
-      File.open(config_file, "w") { |f| f.puts sample_config }
-      puts "Sample config created at #{config_file}"
+      unless File.exist?(config_file)
+        File.open(config_file, "w") { |f| f.puts sample_config }
+        puts "Sample config created at #{config_file}"
+        return false
+      end
 
-      Process.exit 0
+      return true
     end
 
+    ##
+    ## Content for sample configuration
+    ##
+    ## @return     [String] sample config
+    ##
     def sample_config
       <<~EOCONFIG
         tracks:
